Slight renaming.
