--- src/lib/LandingPageWrapper.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	import { Canvas } from '@threlte/core';
-	import LandingPage from '$lib/LandingPage.svelte';
-</script>
-
-<Canvas>
-	<LandingPage />
-</Canvas>
